hello_world/requirements.txt
Flask==1.0.2
Pillow
numpy
opencv-python
